mark plugin: split the data carrying classes into a own module
